packages/tsprint/pyproject.toml
[project]
name = "tsprint"
version = "0.1.0"
description = "PaperCut PaperCut printing client"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests",
    "beautifulsoup4",
    "python-dotenv",
]

[project.scripts]
tsprint = "tsprint.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tsprint"]
